requirements.txt
streamlit
pypdf
google-generativeai
requests
beautifulsoup4
pandas
